+++ fearless/Common/Extension/UIKit/UIStackView+Manage.swift
@@ -1,0 +1,19 @@
+import UIKit
+
+extension UIStackView {
+    func insertArranged(view: UIView, after otherView: UIView) {
+        guard let index = arrangedSubviews.firstIndex(of: otherView) else {
+            return
+        }
+
+        insertArrangedSubview(view, at: index + 1)
+    }
+
+    func insertArranged(view: UIView, befor otherView: UIView) {
+        guard let index = arrangedSubviews.firstIndex(of: otherView) else {
+            return
+        }
+
+        insertArrangedSubview(view, at: index)
+    }
+}
