+++ fearless/Modules/Crowdloan/CrowdloanList/View/ActiveCrowdloanTableViewCell.swift
@@ -1,25 +1,10 @@
 import UIKit
 
-final class ActiveCrowdloanTableViewCell: UITableViewCell {
-    let titleLabel: UILabel = {
-        let label = UILabel()
-        label.textColor = R.color.colorWhite()
-        label.font = .p1Paragraph
-        label.textAlignment = .left
-        return label
-    }()
-
-    let iconImageView: UIImageView = {
-        let imageView = UIImageView()
-        imageView.contentMode = .scaleAspectFit
-        return imageView
-    }()
-
+final class ActiveCrowdloanTableViewCell: BaseCrowdloanTableViewCell {
     let timeLabel: UILabel = {
         let label = UILabel()
         label.textColor = R.color.colorLightGray()
         label.font = .p3Paragraph
-        label.textAlignment = .right
         return label
     }()
 
@@ -29,40 +14,6 @@ final class ActiveCrowdloanTableViewCell: UITableViewCell {
         imageView.contentMode = .center
         return imageView
     }()
-
-    let detailsLabel: UILabel = {
-        let label = UILabel()
-        label.textColor = R.color.colorLightGray()
-        label.font = .p2Paragraph
-        label.lineBreakMode = .byTruncatingMiddle
-        return label
-    }()
-
-    let progressLabel: UILabel = {
-        let label = UILabel()
-        label.textColor = R.color.colorWhite()
-        label.font = .p2Paragraph
-        return label
-    }()
-
-    let mainStackView: UIStackView = {
-        let view = UIStackView()
-        view.axis = .vertical
-        view.distribution = .fill
-        view.alignment = .leading
-        return view
-    }()
-
-    let titleStackView: UIStackView = {
-        let view = UIStackView()
-        view.axis = .horizontal
-        view.distribution = .fill
-        view.alignment = .center
-        view.spacing = 8
-        return view
-    }()
-
-    private(set) var contributionLabel: UILabel?
 
     private var viewModel: ActiveCrowdloanViewModel?
 
@@ -78,13 +29,19 @@ final class ActiveCrowdloanTableViewCell: UITableViewCell {
     override init(style: UITableViewCell.CellStyle, reuseIdentifier: String?) {
         super.init(style: style, reuseIdentifier: reuseIdentifier)
 
-        configure()
-        setupLayout()
+        applyStyle()
     }
 
-    @available(*, unavailable)
-    required init?(coder _: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
+    private func applyStyle() {
+        titleLabel.textColor = R.color.colorWhite()
+        titleLabel.font = .p1Paragraph
+
+        detailsLabel.textColor = R.color.colorLightGray()
+        detailsLabel.font = .p2Paragraph
+        detailsLabel.lineBreakMode = .byTruncatingMiddle
+
+        progressLabel.textColor = R.color.colorWhite()
+        progressLabel.font = .p2Paragraph
     }
 
     func bind(viewModel: ActiveCrowdloanViewModel) {
@@ -105,11 +62,11 @@ final class ActiveCrowdloanTableViewCell: UITableViewCell {
         timeLabel.text = viewModel.timeleft
 
         if let contribution = viewModel.contribution {
-            insertContributionIfNeeded()
+            showContributionLabel()
 
             contributionLabel?.text = contribution
         } else {
-            removeContributionIfNeeded()
+            hideContributionLabel()
         }
 
         viewModel.iconViewModel.loadImage(
@@ -119,83 +76,20 @@ final class ActiveCrowdloanTableViewCell: UITableViewCell {
         )
     }
 
-    private func configure() {
-        backgroundColor = .clear
+    override func setupLayout() {
+        super.setupLayout()
 
-        selectedBackgroundView = UIView()
-        selectedBackgroundView?.backgroundColor = R.color.colorCellSelection()
-
-        separatorInset = UIEdgeInsets(
-            top: 0.0,
-            left: UIConstants.horizontalInset,
-            bottom: 0.0,
-            right: UIConstants.horizontalInset
-        )
-    }
-
-    private func setupLayout() {
-        contentView.addSubview(iconImageView)
-
-        iconImageView.snp.makeConstraints { make in
-            make.leading.equalToSuperview().inset(UIConstants.horizontalInset)
-            make.size.equalTo(CrowdloanViewConstants.iconSize)
-            make.top.equalToSuperview().inset(11)
-        }
-
-        contentView.addSubview(mainStackView)
-
-        mainStackView.snp.makeConstraints { make in
-            make.leading.equalTo(iconImageView.snp.trailing).offset(UIConstants.horizontalInset)
-            make.trailing.equalToSuperview().inset(UIConstants.horizontalInset)
-            make.top.equalToSuperview().inset(6.0)
-            make.bottom.equalToSuperview().inset(12.0)
-        }
-
-        mainStackView.addArrangedSubview(titleStackView)
-
-        titleStackView.snp.makeConstraints { make in
-            make.width.equalToSuperview()
-        }
-
-        titleStackView.addArrangedSubview(titleLabel)
         titleStackView.addArrangedSubview(timeLabel)
         titleStackView.addArrangedSubview(navigationImageView)
 
         navigationImageView.snp.makeConstraints { make in
             make.size.equalTo(24)
         }
-
-        mainStackView.addArrangedSubview(detailsLabel)
-        mainStackView.setCustomSpacing(8.0, after: detailsLabel)
-
-        mainStackView.addArrangedSubview(progressLabel)
     }
 
-    private func insertContributionIfNeeded() {
-        guard contributionLabel == nil else {
-            return
-        }
+    override func showContributionLabel() {
+        super.showContributionLabel()
 
-        let label = UILabel()
-        label.textColor = R.color.colorWhite()
-        label.font = .p2Paragraph
-
-        mainStackView.addArrangedSubview(label)
-        mainStackView.setCustomSpacing(8.0, after: progressLabel)
-
-        contributionLabel = label
-    }
-
-    private func removeContributionIfNeeded() {
-        guard let label = contributionLabel else {
-            return
-        }
-
-        mainStackView.removeArrangedSubview(label)
-        mainStackView.setCustomSpacing(0.0, after: progressLabel)
-
-        label.removeFromSuperview()
-
-        contributionLabel = nil
+        contributionLabel?.textColor = R.color.colorWhite()
     }
 }
